move error logic to ui-validated component
